+++ crates/block-producer/src/server.rs
@@ -1,9 +1,13 @@
 use std::{collections::HashMap, net::SocketAddr, time::Duration};
 
 use anyhow::{Context, Result};
+use futures::StreamExt;
 use miden_node_proto::{
+    domain::mempool::MempoolEvent,
     generated::{
-        block_producer::api_server,
+        block_producer::{
+            MempoolEvent as ProtoMempoolEvent, MempoolSubscriptionRequest, api_server,
+        },
         requests::SubmitProvenTransactionRequest,
         responses::{BlockProducerStatusResponse, SubmitProvenTransactionResponse},
     },
@@ -15,10 +19,11 @@ use miden_node_utils::{
     tracing::grpc::{OtelInterceptor, block_producer_trace_fn},
 };
 use miden_objects::{
-    note::Nullifier, transaction::ProvenTransaction, utils::serde::Deserializable,
+    block::BlockNumber, note::Nullifier, transaction::ProvenTransaction,
+    utils::serde::Deserializable,
 };
 use tokio::{net::TcpListener, sync::Mutex};
-use tokio_stream::wrappers::TcpListenerStream;
+use tokio_stream::wrappers::{ReceiverStream, TcpListenerStream};
 use tonic::Status;
 use tower_http::trace::TraceLayer;
 use tracing::{debug, error, info, instrument};
@@ -235,6 +240,48 @@ impl api_server::Api for BlockProducerRpcServer {
             status: "connected".to_string(),
         }))
     }
+
+    type MempoolSubscriptionStream = MempoolEventSubscription;
+
+    async fn mempool_subscription(
+        &self,
+        request: tonic::Request<MempoolSubscriptionRequest>,
+    ) -> Result<tonic::Response<Self::MempoolSubscriptionStream>, tonic::Status> {
+        let chain_tip = BlockNumber::from(request.into_inner().chain_tip);
+
+        let subscription =
+            self.mempool
+                .lock()
+                .await
+                .lock()
+                .await
+                .subscribe(chain_tip)
+                .map_err(|mempool_tip| {
+                    tonic::Status::invalid_argument(format!(
+                        "Mempool's chain tip {mempool_tip} does not match request's {chain_tip}"
+                    ))
+                })?;
+        let subscription = ReceiverStream::new(subscription);
+
+        Ok(tonic::Response::new(MempoolEventSubscription { inner: subscription }))
+    }
+}
+
+struct MempoolEventSubscription {
+    inner: ReceiverStream<MempoolEvent>,
+}
+
+impl tokio_stream::Stream for MempoolEventSubscription {
+    type Item = Result<ProtoMempoolEvent, tonic::Status>;
+
+    fn poll_next(
+        mut self: std::pin::Pin<&mut Self>,
+        cx: &mut std::task::Context<'_>,
+    ) -> std::task::Poll<Option<Self::Item>> {
+        self.inner
+            .poll_next_unpin(cx)
+            .map(|x| x.map(ProtoMempoolEvent::from).map(Result::Ok))
+    }
 }
 
 impl BlockProducerRpcServer {
@@ -252,11 +299,21 @@ impl BlockProducerRpcServer {
             .build_v1()
             .context("failed to build reflection service")?;
 
+        // This is currently required for postman to work properly because
+        // it doesn't support the new version yet.
+        //
+        // See: <https://github.com/postmanlabs/postman-app-support/issues/13120>.
+        let reflection_service_alpha = tonic_reflection::server::Builder::configure()
+            .register_file_descriptor_set(block_producer_api_descriptor())
+            .build_v1alpha()
+            .context("failed to build reflection service")?;
+
         // Build the gRPC server with the API service and trace layer.
         tonic::transport::Server::builder()
             .layer(TraceLayer::new_for_grpc().make_span_with(block_producer_trace_fn))
             .add_service(api_server::ApiServer::new(self))
             .add_service(reflection_service)
+            .add_service(reflection_service_alpha)
             .serve_with_incoming(TcpListenerStream::new(listener))
             .await
             .context("failed to serve block producer API")