+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 - Replaced integer-based duration args with human-readable duration strings (#998 & #1014).
 - [BREAKING] Refactor the `miden-proving-service` proxy status service to use gRPC instead of HTTP (#953).
 - Added configurable network id for the faucet (#1016).
+- Support gRPC server relection `v1alpha` (#1036).
 
 ### Changes
 
@@ -29,6 +30,10 @@
 - Faucet `PoW` difficulty is now configurable (#924).
 - Removed system monitor from node binary (#1019).
 - [BREAKING] Renamed `open_telemetry` to `enable_otel` in all node's commands (#1019).
+- [BREAKING] Rename `miden-proving-service` to `miden-remote-prover` (#1004).
+- [BREAKING] Rename `miden-proving-service-client` to `miden-remote-prover-client` (#1004).
+- [BREAKING] Rename `RemoteProverError` to `RemoteProverClientError` (#1004).
+- [BREAKING] Rename `ProvingServiceError` to `RemoteProverError` (#1004).
 
 ### Fixes